pyproject.toml
[project]
name = "q-18"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "numpy>=2.4.2",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
